Added Block Interaction script
This script makes possible the placement of the blocks in the world and the ability to remove them. This script has to be added as script component to the player. In the inspector you'll have to drag and drop the camera (inside the player object in the scene), the highlighted material and the block prefab that will be placed in the world.